requirements.txt
# LangGraph 1.0
langgraph==1.0.1

# LangChain 1.0
langchain==1.0.0
langchain-core==1.0.0
langchain-openai==1.0.0

# MCP SDK
mcp==1.1.0

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.31.1

# OpenAI
openai==1.52.0

# PostgreSQL with pgvector
psycopg2-binary==2.9.9
pgvector==0.3.4

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.27.2

# Data Processing
pandas==2.2.3
numpy==2.1.2

# Environment
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Logging
colorlog==6.9.0
